[AIM] Fix legacy SNAT migration
Avoid issues caused by missing mapping DB records when cleaning up the
legacy plugin's SNAT-related resources.

+++ gbpservice/neutron/plugins/ml2plus/drivers/apic_aim/mechanism_driver.py
@@ -803,16 +803,18 @@ class ApicMechanismDriver(api_plus.MechanismDriver,
                 # Before we can clean up the default vrf, we have to
                 # remove the association in the network_mapping first.
                 mapping = self._get_network_mapping(session, current['id'])
-                self._set_network_vrf(mapping, self._map_unrouted_vrf())
-                vrf = self._map_default_vrf(session, current)
-                self._cleanup_default_vrf(aim_ctx, vrf)
+                if mapping:
+                    self._set_network_vrf(mapping, self._map_unrouted_vrf())
+                    vrf = self._map_default_vrf(session, current)
+                    self._cleanup_default_vrf(aim_ctx, vrf)
         else:
             mapping = self._get_network_mapping(session, current['id'])
-            bd = self._get_network_bd(mapping)
-            self.aim.delete(aim_ctx, bd)
-            epg = self._get_network_epg(mapping)
-            self.aim.delete(aim_ctx, epg)
-            session.delete(mapping)
+            if mapping:
+                bd = self._get_network_bd(mapping)
+                self.aim.delete(aim_ctx, bd)
+                epg = self._get_network_epg(mapping)
+                self.aim.delete(aim_ctx, epg)
+                session.delete(mapping)
 
     def extend_network_dict_bulk(self, session, results, single=False):
         # Gather db objects
@@ -1349,6 +1351,12 @@ class ApicMechanismDriver(api_plus.MechanismDriver,
                             l3_db.RouterPort.port_type ==
                             n_constants.DEVICE_OWNER_ROUTER_INTF)):
             ip_address, subnet_db, network_db = intf
+
+            if not network_db.aim_mapping:
+                LOG.warning(
+                    "Mapping missing for network %s in extend_router_dict" %
+                    network_db.id)
+                continue
 
             if network_db.aim_mapping.bd_name:
                 bd = self._get_network_bd(network_db.aim_mapping)
